requirements.txt
langchain
langgraph
openai
pandas
scikit-learn
langchain-openai
tiktoken
pyyaml